requirements.txt
fastapi==0.125.0
uvicorn==0.38.0
pika==1.3.2
transformers==4.52.3
sentence_transformers==5.2.0
torch
python-dotenv
supabase
